requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=1.4.0
pydantic>=1.10.0
passlib[bcrypt]>=1.7.0
pytest>=7.0.0
httpx>=0.24.0
alembic>=1.11.0
python-jose[cryptography] >= 3.5.0

aiokafka~=0.12.0
python-dotenv~=1.1.0
Pyrogram~=2.0.106
pytest-asyncio~=1.0.0